move ShareContentButton to be a child of ShareContentButtonConnected
