+++ packages/apollos-ui-connected/src/ShareContentButtonConnected/ShareContentButton/index.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { compose, pure } from 'recompose';
 import { Touchable, Icon, withTheme } from '@apollosproject/ui-kit';
 import { AnalyticsConsumer } from '@apollosproject/ui-analytics';
-import { share } from '../utils';
+import { share } from '../../utils';
 
 const enhance = compose(
   pure,
