+++ packages/apollos-ui-connected/src/ShareContentButtonConnected/index.js
@@ -1,3 +1,5 @@
 import ShareContentButtonConnected from './ShareContentButtonConnected';
 
+export ShareContentButton from './ShareContentButton';
+
 export default ShareContentButtonConnected;
